# backend/src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "CRIO Xmeme API submission",
    "version": "1.0.0",
    "title": "XMEME",
    "contact": {
      "name": "Yash Agarwal",
      "email": "yagarwal1307@gmail.com"
    },
    "license": {
      "name": "AGPLv3.0",
      "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
    }
  },
  "host": "xmeme1307.herokuapp.com",
  "basePath": "/",
  "tags": [
    {
      "name": "memes",
      "description": "Share memes with community"
    }
  ],
  "schemes": [
    "https"
  ],
  "paths": {
    "/memes": {
      "post": {
        "tags": [
          "memes"
        ],
        "summary": "Create a new meme",
        "description": "Create a new meme",
        "operationId": "addMeme",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Meme object that you want to share",
            "required": true,
            "schema": {
              "$ref": "#/definitions/MemePost"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Meme created successfully",
            "schema": {
              "$ref": "#/definitions/IdResponse"
            }
          },
          "400": {
            "description": "Invalid request body"
          },
          "409": {
            "description": "Meme with same payload already exists"
          }
        }
      },
      "get": {
        "tags": [
          "memes"
        ],
        "summary": "Get memes",
        "description": "Returns 100 latest memes sorted according to the time of creation",
        "operationId": "getMemes",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Memes retrieved successfully",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/MemeResponse"
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/memes/{memeId}": {
      "get": {
        "tags": [
          "memes"
        ],
        "summary": "Find meme by ID",
        "description": "Returns a single meme",
        "operationId": "getMemeById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "memeId",
            "in": "path",
            "description": "ID of meme to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "retrieved meme successfully",
            "schema": {
              "$ref": "#/definitions/MemeResponse"
            }
          },
          "404": {
            "description": "Meme not found"
          },
          "422": {
            "description": "Invalid Meme object ID"
          },
          "500": {
            "description": "Inernal Server error"
          }
        }
      },
      "patch": {
        "tags": [
          "memes"
        ],
        "summary": "Updates meme by id",
        "description": "Update the caption or URL of an individual meme",
        "operationId": "updateMeme",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "memeId",
            "in": "path",
            "description": "Meme id to update",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "New meme object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/MemePatch"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Meme successfully updated"
          },
          "400": {
            "description": "Invalid Payload"
          },
          "404": {
            "description": "Meme not found"
          },
          "409": {
            "description": "Meme with same payload already exists"
          },
          "422": {
            "description": "Invalid meme object ID"
          }
        }
      }
    }
  },
  "definitions": {
    "MemeResponse": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Yash Agarwal",
          "description": "Minimum length is 1"
        },
        "url": {
          "type": "string",
          "example": "https://static.mommypoppins.com/styles/image620x420/s3/school_meme_3_1.jpg?itok=zJWEcZFx",
          "description": "Should be a valid image URL"
        },
        "caption": {
          "type": "string",
          "example": "Some random caption",
          "description": "Minimum length is 1"
        },
        "id": {
          "type": "string"
        }
      }
    },
    "MemePost": {
      "type": "object",
      "required": [
        "name",
        "caption",
        "url"
      ],
      "properties": {
        "name": {
          "type": "string",
          "example": "Yash Agarwal",
          "description": "Minimum length is 1"
        },
        "url": {
          "type": "string",
          "example": "https://static.mommypoppins.com/styles/image620x420/s3/school_meme_3_1.jpg?itok=zJWEcZFx",
          "description": "Should be a valid image URL"
        },
        "caption": {
          "type": "string",
          "example": "Some random caption",
          "description": "Minimum length is 1"
        }
      }
    },
    "MemePatch": {
      "type": "object",
      "properties": {
        "url": {
          "type": "string",
          "example": "https://static.mommypoppins.com/styles/image620x420/s3/school_meme_3_1.jpg?itok=zJWEcZFx",
          "description": "Should be a valid image URL"
        },
        "caption": {
          "type": "string",
          "example": "Some random caption",
          "description": "Minimum length is 1"
        }
      }
    },
    "IdResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "description": "A unique object id auto-generated by mongodb"
        }
      }
    }
  }
}